[MISC] move type_traits/pre to utility
