+++ include/seqan3/range/views/interleave.hpp
@@ -13,18 +13,18 @@
 #pragma once
 
 #include <cmath>
+#include <seqan3/std/concepts>
+#include <seqan3/std/ranges>
 
 #include <range/v3/view/chunk.hpp>
 
-#include <seqan3/core/type_traits/pre.hpp>
 #include <seqan3/core/type_traits/transformation_trait_or.hpp>
 #include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/views/join.hpp>
 #include <seqan3/range/views/persist.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
-#include <seqan3/std/concepts>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/type_traits/pre.hpp>
 
 namespace seqan3::detail
 {
